authorization though passport in cat endpoints

+++ week_two/models/catModel.js
@@ -37,11 +37,11 @@ const addCat = async (cat, res) => {
   }
 };
 
-const deleteCatById = async (res, catId) => {
+const deleteCatById = async (res, owner, catId) => {
   try {
     const [rows] = await promisePool.query(
-      "DELETE FROM wop_cat WHERE cat_id = ?",
-      [catId]
+      "DELETE FROM wop_cat WHERE cat_id = ? AND owner = ?",
+      [catId, owner]
     );
     return rows[0];
   } catch (e) {
